[修改] 支持子结构

+++ ExeclTool/gulpfile.ts
@@ -141,7 +141,10 @@ gulp.task("copyOutput",(cb)=>{
         }else{
             fs.mkdirSync(outpath);
         }
-        fs.cpSync(path.join(tmp_path,key), outpath, { recursive: true });
+        if(fs.existsSync(path.join(tmp_path,key))) {
+            fs.cpSync(path.join(tmp_path,key), outpath, { recursive: true });
+        }
+        
     }
     cb();
 });
